renamed and reorganized surf gifti files
Previous naming was not really bids-compliant because of the .native.
and .unfolded.

This is solved now by removing the .native. and .unfolded. and just
using the space- tag, with a new space-unfolded to refer to unfolded
data.

specifically:
native and space-corobl are now just referred to as space-corobl.
native and space-T1w are now just space-T1w.
unfolded and space-corobl are now just space-unfolded.